% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_nest_discrete.R
\name{scale_nest_manual}
\alias{scale_nest_manual}
\title{Scales NEST manually or some such thing.}
\usage{
scale_nest_manual(
  aesthetics,
  scale_name,
  palette = stream_pallette,
  name = waiver(),
  breaks = waiver(),
  labels = waiver(),
  limits = NULL,
  expand = waiver(),
  na.translate = TRUE,
  na.value = NA,
  drop = TRUE,
  guide = "legend",
  position = "left",
  super = ScaleDiscrete
)
}
\arguments{
\item{palette}{(``)}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Adds scaling manually or some such thing. IDK
what this does TBH, I'm just trying to
appease R CMD check.
}
