% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nest.R
\name{theme_nest}
\alias{theme_nest}
\title{Returns a custom \code{NEST} \code{ggplot2} theme}
\usage{
theme_nest(font_size = 10, ...)
}
\arguments{
\item{font_size}{(\code{numeric}) font size to be passed to \verb{[ggplot2::element_text()]}}

\item{...}{(\code{list}) additional arguments to \verb{[ggplot2::theme()]}}
}
\value{
Return ggplot theme
}
\description{
Returns a custom \code{NEST} \code{ggplot2} theme
}
\examples{
plot <- ggplot2::ggplot(iris, ggplot2::aes(x = Sepal.Length, y = Sepal.Width)) +
  ggplot2::geom_point() +
  theme_nest()
}
